Send a UDP packet to the Gateway-IP every second
This is intended to help keeping the connection maintained better
Sometimes for a LYT8266 on the fringe of WiFi signal, a lost connection is otherwise not detected early
In this case LYT8266 is mounted in a lamp with a lot of RF-attenuating metal around and WiFi signal is poor due to this. Hopefully this is an alternative to a dedicated repeater.